Clean up, preparing watch
