+++ dev/package.json
@@ -2,10 +2,6 @@
   "name": "webpack-focus-dev",
   "version": "0.11.5",
   "description": "Focus webpack config (Development)",
-  "main": "lib/index.js",
-  "bin": {
-    "focus-dev-server": "./lib/launcher.js"
-  },
   "scripts": {},
   "repository": {
     "type": "git",
@@ -18,23 +14,6 @@
   "author": "focus@kleegroup.com",
   "license": "MIT",
   "homepage": "https://github.com/KleeGroup/webpack-focus#readme",
-  "betterScripts": {
-    "webpack-dev-server-npm-focus": {
-      "command": "./dev-server.js",
-      "env": {
-        "ANCHOR_CLASS": "demo-app",
-        "BABELIFIED_PATH": "./app",
-        "ENTRY_FILE_PATH": "./app/index",
-        "PAGE_TITLE": "Focus demonstration application",
-        "API_HOST": "localhost",
-        "API_PORT": 8080,
-        "OUTPUT_DIR": "./public",
-        "LOCAL_FOCUS": false,
-        "LEGACY_SEARCH_API": false,
-        "BASE_URL": "/"
-      }
-    }
-  },
   "dependencies": {},
   "devDependencies": {},
   "peerDependencies": {}
