+++ package.json
@@ -10,7 +10,8 @@
     "builder:gen-dev": "builder-support gen-dev",
     "build": "babel *.js --out-dir lib",
     "prepublish": "npm run build",
-    "focus-dev-server": "focus-dev-server"
+    "focus-dev-server": "focus-dev-server",
+    "focus-webpack": "webpack"
   },
   "repository": {
     "type": "git",
@@ -23,33 +24,13 @@
   "author": "focus@kleegroup.com",
   "license": "MIT",
   "homepage": "https://github.com/KleeGroup/webpack-focus#readme",
-  "betterScripts": {
-    "webpack-dev-server-npm-focus": {
-      "command": "./dev-server.js",
-      "env": {
-        "ANCHOR_CLASS": "demo-app",
-        "BABELIFIED_PATH": "./app",
-        "ENTRY_FILE_PATH": "./app/index",
-        "PAGE_TITLE": "Focus demonstration application",
-        "API_HOST": "localhost",
-        "API_PORT": 8080,
-        "OUTPUT_DIR": "./public",
-        "LOCAL_FOCUS": false,
-        "LEGACY_SEARCH_API": false,
-        "BASE_URL": "/"
-      }
-    }
-  },
   "dependencies": {
     "autoprefixer-loader": "^3.2.0",
     "babel-cli": "^6.3.17",
     "babel-core": "^6.3.17",
     "babel-loader": "^6.2.10",
     "babel-plugin-replace-require": "0.0.4",
-    "babel-preset-es2015": "^6.3.13",
     "babel-preset-focus": "file:../babel-preset-focus",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-stage-0": "^6.3.13",
     "css-loader": "^0.26.1",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
